% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_usc}
\alias{scale_color_usc}
\title{Color scale for usc}
\usage{
scale_color_usc(palette = "simple", discrete = TRUE, rev = FALSE, ...)
}
\arguments{
\item{palette}{Character name for palettes}

\item{discrete}{boolean indicating data type}

\item{rev}{boolean to reverse palette}

\item{...}{additional arguements for discrete_scale() or scale_color_gradientn()}
}
\description{
See official color guide \url{https://sc.edu/about/offices_and_divisions/communications/toolbox/visuals/colors/using_color/index.php}
}
