% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_fill_uofsc}
\alias{scale_fill_uofsc}
\title{Fill scale for UofSC}
\usage{
scale_fill_uofsc(palette = "simple", discrete = TRUE, rev = FALSE, ...)
}
\arguments{
\item{palette}{Character name for palettes}

\item{discrete}{boolean indicating data type}

\item{rev}{boolean to reverse palette}

\item{...}{additional arguements for discrete_scale() or scale_color_gradientn()}
}
\description{
Fill scale for UofSC
}
