% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_scales.R
\name{uofsc_cols}
\alias{uofsc_cols}
\title{Access UofSC colors}
\usage{
uofsc_cols(...)
}
\arguments{
\item{...}{character names of uofsc colors}
}
\description{
A function to extract official UofSC colors
}
\references{
\url{https://sc.edu/about/offices_and_divisions/communications/toolbox/visuals/colors/}
}
