% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_scales.R
\name{usc_cols}
\alias{usc_cols}
\title{Access usc colors}
\usage{
usc_cols(...)
}
\arguments{
\item{...}{character names of usc colors}
}
\description{
A function to extract official usc colors
}
\references{
\url{https://sc.edu/about/offices_and_divisions/communications/toolbox/visuals/colors/}
}
