% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_scales.R
\name{usc_pal}
\alias{usc_pal}
\title{Palette interpolator}
\usage{
usc_pal(palette = "simple", rev = FALSE, ...)
}
\arguments{
\item{palette}{char name of palette}

\item{rev}{boolean ot reverse palette}

\item{...}{additional to pass to colorRampPalette}
}
\description{
This will interpolate between colors
}
